Compliance com modelo MVC e Clean Code
Juntei classes redundantes e criei um enum pra melhorar leitura.
